requirements.txt
llama-index
langchain
guardrails-ai
openai
qdrant-client
python-dotenv
